fix(scroll): Deliver touch and wheel deltas unquantized
Manipulation deltas were held back until they crossed a significance
threshold, so a slow drag advanced the content in 2px steps instead of
following the finger, and wheel deltas were truncated to whole detents by an
int division.

Scrolling asks for every delta as it arrives; the threshold exists to suppress
accidental micro-manipulations, which is not what a scroll presenter wants.

+++ src/Uno.UI.Composition/Composition/InteractionTracker/InteractionTrackerCustomAnimationState.cs
@@ -35,7 +35,7 @@ internal sealed class InteractionTrackerCustomAnimationState : InteractionTracke
 	{
 	}
 
-	internal override void ReceivePointerWheel(int delta, bool isHorizontal)
+	internal override void ReceivePointerWheel(double delta, bool isHorizontal)
 	{
 	}
